Separate login logic from login page
Login page no longer contains the logic for making login calls to the api.
